foodfor/foodfor/requirements.txt
--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/

pinax-theme-bootstrap==2.2.2
pinax-theme-bootstrap-account==1.0b2
django-user-accounts==1.0b4
django-forms-bootstrap==2.0.3.post1
metron==1.0
pinax-utils==1.0b1.dev3
